QStyle: let styleCachePixmap() use the correct devicePixelRatio
The helper function styleCachePixmap() might not be using the correct
dpr. It was written in a time when there was only one application-wide
dpr available but nowadays we have to use the dpr from the paintDevice
we paint on instead.
As a drive-by replace the one usage of styleCacheImage() by
styleCachePixmap() and remove the now superfluous function.

Change-Id: I14eb3fd3bad7f9299d42b63b03fa1f68b4f81cd6
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Wladimir Leuschner <wladimir.leuschner@qt.io>