Enable framework builds for all Apple platforms, including static builds
Static frameworks are now a supported by Apple, so we switch the default
for all Apple platforms, including Qt for iOS and visionOS.

 https://developer.apple.com/documentation/xcode/creating-a-static-framework

This allows us to bundle resources with our libraries, such as the
privacy manifest.

Xcode 15 and later will ensure that the linked and embedded framework
is not copied into the application bundle, as it's already linked into
the app as a static library.

Change-Id: Iaf64e7df2d6969d9562fb31ba65f05767e14506e
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>