QDataStream::readBytes: do not leak the memory in case of bad_alloc
The algorithm was leaking the previous buffer if the attempt to
allocate a new one failed with std::bad_alloc.

Wrap the buffer allocations into std::unique_ptr, so that we do
not leak the buffer in case of bad_alloc.

This patch is not changing the behavior of QDataStream, so it's
NOT handling the bad_alloc exception, and also does not change
the state of the buffer in case the exception occurs.

Change-Id: I5558a6d03768094e4ee83ca47cacf4e09c36f1cf
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
(cherry picked from commit e176dd78fd2f253eb2625585b2bd90b5713e5984)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 06bb5f19bfe686cb149150a2c1a94699ea3e1413)