QDataStream::readBytes: do not leak the memory in case of bad_alloc
The algorithm was leaking the previous buffer if the attempt to
allocate a new one failed with std::bad_alloc.

Wrap the buffer allocations into std::unique_ptr, so that we do
not leak the buffer in case of bad_alloc.

This patch is not changing the behavior of QDataStream, so it's
NOT handling the bad_alloc exception, and also does not change
the state of the buffer in case the exception occurs.

Pick-to: 6.7 6.5
Change-Id: I5558a6d03768094e4ee83ca47cacf4e09c36f1cf
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>