Android: Delete parent window in QtView instead of just the child view
Since the parent window wraps the QtView, it should always be destroyed
when the QtView is, and live inside QtView rather than the delegate.

Destroying the parent window will always destroy the child window, so
do not destroy the child window separately.

Move createRootWindow and deleteWindow native functions to QtView.

Fixes: QTBUG-124908
Change-Id: Ib6b3c6388a9dd3f74d16fa09a442b0a6f8ccb336
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>