QThread: replace three booleans with one state variable
There aren't 2^3 possible states for {running, isInFinish, finished}.
There are only three possible states for them.

There's a fourth case of NotStarted, which in my opinion is a design
flaw. It needs to exist so QThread().isFinished() is false, but in a
green field design we would be the same as QProcess: once a process/
thread finishes, it goes back to NotRunning.

I also made the Windows version set back to NotStarted when failing to
start, matching the Unix version.

Drive-by use NSDMI.

Change-Id: I262c3499666e4f4fbcfbfffd17cb3a48dad045dc
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>